cpu-node/requirements.txt
fastapi
uvicorn
requests
pydantic
numpy
faiss-cpu
pdfplumber
sentence-transformers
